feat: 게임 회원 투표를 훅으로 분리
