+++ src/components/molecules/BalanceGameBox/BalanceGameBox.tsx
@@ -2,12 +2,10 @@ import React, { useState } from 'react';
 import { GameOption } from '@/types/game';
 import { getRandomNumbers } from '@/utils/calculator';
 import BalanceGameButton from '@/components/atoms/BalanceGameButton/BalanceGameButton';
-import { useCreateGameVoteMutation } from '@/hooks/api/vote/useCreateGameVoteMutation';
-import { useEditGameVoteMutation } from '@/hooks/api/vote/useEditGameVoteMutation';
-import { useDeleteGameVoteMutation } from '@/hooks/api/vote/useDeleteGameVoteMutation';
 import { useNewSelector } from '@/store';
 import { selectAccessToken } from '@/store/auth';
 import { MyVoteOption, VoteOption } from '@/types/vote';
+import { useUserGameVote } from '@/hooks/game/useBalanceGameVote';
 import * as S from './BalanceGameBox.style';
 
 export interface BalanceGameBoxProps {
@@ -44,36 +42,11 @@ const BalanceGameBox = ({
   const [backgroundImages] = useState<string[]>(getRandomImages);
   const [backgroundImageA, backgroundImageB] = backgroundImages;
 
-  const { mutate: createGameVote } = useCreateGameVoteMutation(
+  const { handleUserGameVote } = useUserGameVote(
     gameSetId,
     gameId,
+    handleNextStage,
   );
-  const { mutate: editGameVote } = useEditGameVoteMutation(gameSetId, gameId);
-  const { mutate: deleteGameVote } = useDeleteGameVoteMutation(
-    gameSetId,
-    gameId,
-  );
-
-  const handleUserGameVote = (
-    selectedOption: MyVoteOption,
-    voteOption: VoteOption,
-  ) => {
-    if (!selectedOption) {
-      createGameVote(voteOption, {
-        onSuccess: () => {
-          const nextStageTimer = setTimeout(() => {
-            handleNextStage();
-          }, 500);
-
-          return () => clearTimeout(nextStageTimer);
-        },
-      });
-    } else if (selectedOption === voteOption) {
-      deleteGameVote();
-    } else {
-      editGameVote(voteOption);
-    }
-  };
 
   const handleButtonClick = (voteOption: VoteOption) => {
     if (accessToken) {
